Moved Sn_scale_estimator from src/dwi/ to lib/math/
