+++ src/dwi/noise_estimator.h
@@ -20,7 +20,7 @@
 #include "dwi/gradient.h"
 #include "math/least_squares.h"
 #include "math/SH.h"
-#include "dwi/Sn_scale_estimator.h"
+#include "math/Sn_scale_estimator.h"
 
 
 namespace MR {
@@ -63,7 +63,7 @@ namespace MR {
             OutputImageType noise;
             Eigen::MatrixXd H, S, R;
             Eigen::VectorXd leverage;
-            Sn_scale_estimator<default_type> scale_estimator;
+            Math::Sn_scale_estimator<default_type> scale_estimator;
             int axis;
         };
     }
